requirements.txt
msgpack
python-dateutil
pyyaml
toolz